add a new seco command (`deploy-debug-pod`) for better debugging
It's task is to sleep for a given amount of time (1 hout by default)
so that we can `kubectl exec` into in and poke around without
messing up with the actual secobox.

Funny that I really needed the other day and crated a toy example
and came across Isaac's issue just by chance. I guess our needs
align as we use the infra more and more every day :)

Resolves #54.